refactor: 움직임 결과 객체 패키지 변경
