+++ src/main/java/step5/domain/Cars.java
@@ -1,7 +1,7 @@
 package step5.domain;
 
-import step5.view.result.CarMovementResult;
-import step5.view.result.CarMovementRoundResult;
+import step5.domain.result.CarMovementResult;
+import step5.domain.result.CarMovementRoundResult;
 
 import java.util.ArrayList;
 import java.util.List;
