+++ src/main/java/step5/domain/result/CarMovementResult.java
@@ -1,4 +1,4 @@
-package step5.view.result;
+package step5.domain.result;
 
 public class CarMovementResult {
     private final String carName;
